chore: add submenu
