+++ src/components/Menu/Menu.tsx
@@ -1,36 +1,14 @@
-import React from 'react';
-import {StyleSheet} from 'react-native';
+import React, {useState} from 'react';
 
-import {Menu as PaperMenu} from 'react-native-paper';
+import {
+  Menu as PaperMenu,
+  MenuProps as PaperMenuProps,
+} from 'react-native-paper';
 
 import {useTheme} from '../../hooks';
 
-import {MenuItem} from './MenuItem';
-import type {MenuProps} from './types';
-
-import type {Theme} from '../../utils/types';
-
-export const createStyles = (theme: Theme) =>
-  StyleSheet.create({
-    menu: {
-      minWidth: 200,
-      borderRadius: 13,
-      shadowColor: 'rgba(0, 0, 0, 0.05)',
-      shadowRadius: 70,
-      shadowOffset: {width: 0, height: 0},
-      elevation: 5,
-    },
-    content: {
-      paddingTop: 6,
-      paddingBottom: 6,
-      backgroundColor: theme.colors.menuBackground,
-    },
-    groupSeparator: {
-      height: 6,
-      flexShrink: 0,
-      backgroundColor: 'transparent',
-    },
-  });
+import {createStyles} from './styles';
+import {MenuItem, MenuItemProps} from './MenuItem';
 
 const GroupSeparator = () => {
   const theme = useTheme();
@@ -47,19 +25,38 @@ const GroupSeparator = () => {
   );
 };
 
+export interface MenuProps extends Omit<PaperMenuProps, 'anchor' | 'theme'> {
+  anchor: React.ReactNode;
+}
+
 export const Menu: React.FC<MenuProps> & {
   Item: typeof MenuItem;
   GroupSeparator: typeof GroupSeparator;
 } = ({anchor, children, style, ...menuProps}) => {
   const theme = useTheme();
   const styles = createStyles(theme);
+  const [hasActiveSubmenu, setHasActiveSubmenu] = useState(false);
+
+  const handleSubmenuOpen = () => setHasActiveSubmenu(true);
+  const handleSubmenuClose = () => setHasActiveSubmenu(false);
+
   return (
     <PaperMenu
       {...menuProps}
       anchor={anchor}
-      style={[styles.menu, style]}
-      contentStyle={styles.content}>
-      {children}
+      style={[styles.menu, hasActiveSubmenu && styles.menuWithSubmenu, style]}
+      contentStyle={[
+        styles.content,
+        hasActiveSubmenu && styles.contentWithSubmenu,
+      ]}>
+      {React.Children.map(children, child =>
+        React.isValidElement<MenuItemProps>(child)
+          ? React.cloneElement(child, {
+              onSubmenuOpen: handleSubmenuOpen,
+              onSubmenuClose: handleSubmenuClose,
+            })
+          : child,
+      )}
     </PaperMenu>
   );
 };
