+++ src/components/Menu/MenuItem/index.ts
@@ -1,2 +1,1 @@
-export * from './Menu';
 export * from './MenuItem';
