+++ src/components/Menu/SubMenu/SubMenu.tsx
@@ -1,8 +1,10 @@
 import React from 'react';
-import {StyleSheet} from 'react-native';
+
 import {Menu as PaperMenu} from 'react-native-paper';
-import {useTheme} from '../../hooks';
-import {Theme} from '../../utils/types';
+
+import {useTheme} from '../../../hooks';
+
+import {createStyles} from './styles';
 
 interface SubMenuProps {
   visible: boolean;
@@ -31,16 +33,3 @@ export const SubMenu: React.FC<SubMenuProps> = ({
     </PaperMenu>
   );
 };
-
-const createStyles = (theme: Theme) =>
-  StyleSheet.create({
-    menu: {
-      minWidth: 220,
-      marginTop: 0, // Remove negative margin
-      marginLeft: 0, // Remove indent
-    },
-    content: {
-      paddingVertical: 6,
-      backgroundColor: theme.colors.menuBackground,
-    },
-  });
